e2e/components: decouple RemoteHostAgent configuration from Pulumi
Add Configure(t, opts...), SetBaseOptions, and SetComponents to RemoteHostAgent
so agent config can be written via SSH without re-running Pulumi. Configure merges
new options on top of the stored baseline, builds the full datadog.yaml (with
fakeintake URLs, API key, and user config), writes it via SSH, and restarts the
agent with OS-aware service management (systemctl / initctl / service fallback).

Host.Init now wires SetComponents automatically so Configure has access to the
host (SSH) and fakeintake (URLs) from the start.

SetAgentConfig is kept as a thin shim for backward compatibility but now
delegates to Agent.Configure instead of duplicating the logic.
